Change type of some constants to unsigned short

+++ link/.local/include/angle-conv-utils.hpp
@@ -15,31 +15,31 @@
 #include <concepts>
 
 /// milliradians per radian
-inline constexpr unsigned int mrad_per_rad = 1'000;
+inline constexpr unsigned short mrad_per_rad = 1'000;
 
 /// radians per revolution
 template <std::floating_point T>
 inline constexpr T rad_per_rev = M_2_pi<T>;
 
 /// degrees per revolution
-inline constexpr unsigned int deg_per_rev = 360;
+inline constexpr unsigned short deg_per_rev = 360;
 
 /// arcminutes per degree
-inline constexpr unsigned int arcmin_per_deg = 60;
+inline constexpr unsigned short arcmin_per_deg = 60;
 
 /// arcseconds per arcminute
-inline constexpr unsigned int arcsec_per_arcmin = 60;
+inline constexpr unsigned short arcsec_per_arcmin = 60;
 
 /// degrees per radian
 template <std::floating_point T>
 inline constexpr T deg_per_rad = 180 / M_pi<T>;
 
-inline constexpr unsigned int quadrants_per_rev = 4;
-inline constexpr unsigned int sextants_per_rev = 6;
-inline constexpr unsigned int octants_per_rev = 8;
-inline constexpr unsigned int hexacontades_per_rev = 60;
-inline constexpr unsigned int binary_degrees_per_rev = 256;
-inline constexpr unsigned int gradians_per_rev = 400;
+inline constexpr unsigned short quadrants_per_rev = 4;
+inline constexpr unsigned short sextants_per_rev = 6;
+inline constexpr unsigned short octants_per_rev = 8;
+inline constexpr unsigned short hexacontades_per_rev = 60;
+inline constexpr unsigned short binary_degrees_per_rev = 256;
+inline constexpr unsigned short gradians_per_rev = 400;
 
 /// convert to radians from milliradians
 template <std::floating_point T>
